requirements.txt
google-genai>=0.5.0
pathlib>=1.0.1
click>=8.1.0
colorama>=0.4.6
rich>=13.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
